first pass at a docker build for alpine
