+++ alpine.yml
@@ -1,0 +1,19 @@
+version: '3.2'
+
+services:
+   imagebuild:
+     build: 
+       context: .
+       dockerfile: Dockerfile
+       args:
+         IMAGE_DISTRO: alpine
+         IMAGE_VERSION: latest
+         UPDATE_CMD_PART1: mkdir -p /var/cache/apk
+         UPDATE_CMD_PART2: ln -s /var/cache/apk /etc/apk/cache
+         UPDATE_CMD_PART3: apk update
+         INSTALL_CMD: apk add bash sudo dropbear mesa-gl xvfb x11vnc xsetroot xterm twm expect shadow
+         CLEANUP_CMD: apk cache clean
+         RUN_CMD: /input/alpine.sh
+     volumes:
+       - ./output:/output
+       - ./input:/input
