--- assets/all/assets.txt
@@ -1,9 +1,0 @@
-addNonRootUser.sh 1549666397 ddbc436155e4665933c9d688bab0c015
-extractFilesystem.sh 1549660050 3f9295ded6fd1d01fb8eabbc2c6d014c
-ld.so.preload 1548782331 43618bdfa97456206c3f21bd8fc2a9cb
-nosudo 1548782331 6c77f57a4be314c57682240754d85327
-startSSHServer.sh 1548782331 1455ac210cfa0966e3921d68bd4957d7
-startVNCServer.sh 1548782331 a52afb25bd9b9ef7ea87c6de3d88b3ad
-startVNCServerStep2.sh 1549915075 bb459db7e478c016ad14bf8a382e4475
-startXSDLServer.sh 1549918608 e044b2cddb0074da74b983fc94ca0c8c
-userland_profile.sh 1548782331 8c185e7210b2af78d1f6135636737cde
