--- assets/all/ld.so.preload
@@ -1,1 +1,0 @@
-/support/libdisableselinux.so
