--- assets/arm/assets.txt
@@ -1,4 +1,0 @@
-busybox 1549916696 1c81d7e732fd330d11cdbceb2f26b660
-libdisableselinux.so 1549916696 2d50989f0f18bff8b4d9864174f7ea0a
-rootfs.tar.gz.part00 1549916696 352edfa6933fada95fc8c8b4ef905ba7
-rootfs.tar.gz.part01 1549916696 4f4d81650e2ed31ae5fa2299390ddcba
