--- assets/arm64/assets.txt
@@ -1,4 +1,0 @@
-busybox 1549915004 1704c060ac2623cabff43e458dabb60e
-libdisableselinux.so 1549915004 3e4129f41148539fbd7e8bc71ef123f1
-rootfs.tar.gz.part00 1549915004 dcad7280ef90b0bb3c1c5106bea2ab29
-rootfs.tar.gz.part01 1549915004 1bb7a8c25b8216cc7b8d2239fdd16e46
