+++ assets/userland_profile.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+unset LD_PRELOAD
+unset LD_LIBRARY_PATH
+export LIBGL_ALWAYS_SOFTWARE=1
+export PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games
