--- assets/x86/assets.txt
@@ -1,4 +1,0 @@
-busybox 1549916701 75cb4a7a8321fd7550007aef981f8f4e
-libdisableselinux.so 1549916701 b6a6024cda3a303c15f06df0429fde50
-rootfs.tar.gz.part00 1549916700 25e086e5a39959623db7f64e64ebc372
-rootfs.tar.gz.part01 1549916701 d298f328f96db25f06da0aa534212513
