--- assets/x86_64/assets.txt
@@ -1,4 +1,0 @@
-busybox 1549916705 075b73bc6cb873cc1e6a3d97219f682a
-libdisableselinux.so 1549916705 4a7f72d752e2f38a1328563b8ee477f8
-rootfs.tar.gz.part00 1549916705 704fee9754a50c55a1f238e1c8aa7de0
-rootfs.tar.gz.part01 1549916705 54de7593915d43381299bb6b74073d4e
