--- scripts/shrinkRootfs.sh
@@ -1,3 +1,0 @@
-#! /bin/bash
-
-apk cache clean
